+++ src/app/(dashboard)/facteurs-d-emission/[id]/modifier/page.tsx
@@ -3,7 +3,7 @@ import EditEmissionFactorPage from '@/components/pages/EditEmissionFactor'
 import NotFound from '@/components/pages/NotFound'
 import { getOrganizationVersionById } from '@/db/organization'
 import { canEditEmissionFactor } from '@/services/permissions/emissionFactor'
-import { hasAccessToEmissionFactors } from '@/services/permissions/environmentExtended'
+import { hasAccessToEmissionFactors } from '@/services/permissions/environmentAdvanced'
 import { getDetailedEmissionFactor, getEmissionFactorLocations } from '@/services/serverFunctions/emissionFactor'
 import { hasActiveLicence } from '@/utils/organization'
 import { UserSession } from 'next-auth'
